Refactor service handling, replace 'acme_tiny__cert_type'
There are now two independant configuration values for the service
'acme_tiny__service' and the certificate format 'acme_tiny__cert_format'.
There is also a service map which can be overwritten by the user, where
the format for some default services is queried.

Additionally, allow to disable symlinks and restart.